Rename hash/hash_test -> test.ts
